webapp/backend/requirements.txt
# Web App Dependencies (Backend)
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
redis==5.0.1
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9  # PostgreSQL (optional)
aiosqlite==0.19.0       # SQLite async

# HTTP and API
httpx==0.25.2
requests==2.31.0

# File handling and media
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.21.0
imageio==2.9.0
imageio-ffmpeg==0.4.5

# Background tasks
apscheduler==3.10.4

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0